Orchestrator service, permission system, and service heartbeat infrastructure (#67)
* Update and consolidate all docs around networking, orchestration, etc

* Add research on orchestrator SDKs to be used for deploy management

* Update docs / remove incorrect speculation from audit

* Fix Dapr event subscriptions and expand orchestrator service

Dapr Event Subscriptions:
- Add PermissionsEventsController for Dapr pubsub event handling
- Fix CloudEvents parsing by reading request body directly via StreamReader
  instead of [FromBody] JsonElement which failed with Dapr's invoke pattern
- Add CloudEvents format detection (extract data property when present)
- Both DaprEventSubscription and SessionStateChangeEvent tests now pass

Makefile Improvements:
- Add --no-cache to all test build commands (test-http, test-edge,
  test-infrastructure) to prevent Docker layer caching issues during testing

Orchestrator Service:
- Expand schema with comprehensive deployment management APIs
- Add backend orchestrators (Docker Compose, Swarm, Kubernetes, Portainer)
- Add deployment presets system for predefined service configurations
- Add interfaces for all orchestrator components (IOrchestratorRedisManager, etc.)
- Implement deployment events via RabbitMQ pub/sub

Permissions Service:
- Add event handlers for service registration and session state changes
- Integrate with Dapr pubsub for real-time permission updates

Testing:
- Add PermissionsTestHandler with Dapr event subscription tests
- Add OrchestratorTestHandler for deployment lifecycle testing
- Add RabbitMQ wait script and custom Dapr sidecar for reliable event delivery
- All 26 HTTP integration tests now pass



* Fix edge tester port configuration and add Dapr warmup

- Change OpenResty ports from 80/443 to 9080/9443 to avoid conflicts with
  local NGINX (80/443) and bannou service (8080/8443)
- Add service readiness wait to edge-tester that properly verifies Dapr
  sidecar is ready by making a warmup registration call
- Add OpenResty_Host and OpenResty_Port configuration properties
- Fix nginx.conf auth service routing comments for clarity
- Add statestore readiness check to http-tester to fix flaky Accounts tests

The warmup registration call is necessary because login with empty
credentials fails at the validation layer before touching Dapr,
which means Dapr could still be unavailable when tests start.

The statestore readiness check ensures Redis is connected before
running Accounts service tests that depend on Dapr state management.



* Fix WebSocket service routing and add service heartbeat system

Root cause: DaprClient.CreateInvokeMethodRequest strips the /v1.0/invoke/{appId}/method/
prefix before forwarding to the app, but NSwag-generated controllers expect that prefix.
Solution: Use direct HTTP calls to Dapr endpoint (like generated clients do) to preserve
the full path for proper controller route matching.

Key changes:
- Fix ConnectService.RouteToServiceAsync to use direct HTTP instead of DaprClient SDK
- Add ServiceHeartbeatManager for publishing instance health via Dapr pub/sub
- Defer permission registration until after Dapr connectivity confirmed
- Add orchestrator WebSocket edge tests (4 tests now passing)
- Update ServiceHeartbeatEvent schema for multi-service instance reporting
- Add admin role configuration for orchestrator API test authentication
- Clean up test infrastructure and improve test reliability



* Remove extra log files

* 1.0 release cleanup: fix all build warnings and update stub services

Warning fixes:
- Upgrade KubernetesClient 15.0.1 → 18.0.5 (security vulnerability NU1902)
- Fix nullable schema properties (capacity, issues, error) in common-events.yaml/orchestrator-api.yaml
- Fix CS8620 nullability in PermissionsServiceTests with proper Moq patterns
- Fix CS0618 obsolete RedisChannel with RedisChannel.Literal()
- Fix CS0675 bitwise-or in GuidGenerator with uint cast
- Fix CS8601 null reference in ConnectService with explicit null check
- Update KubernetesOrchestrator for v18 API changes (object initializers)

Stub service documentation:
- Update WebsiteService, BehaviorService, GameSessionService XML docs to "Not yet implemented"
- Update corresponding test file comments to reflect planned-for-future status

Code cleanup:
- Remove duplicate ConnectionState definition from Program.cs (kept in ConnectionState.cs)
- Remove obsolete HandleClientConnectionAsync placeholder

Result: 0 warnings, 0 errors, 326 tests passing



* Upgrade RabbitMQ 3.12 → 4.0 (EOL version fix)

RabbitMQ 3.12 reached end of life in March 2024. Upgraded to 4.0-management
which is the current stable LTS release with community support.

Verified: 6/6 infrastructure tests pass, 69/69 HTTP tests pass



* Try fixing forward compatibility edge test step in CI, docker hostname resolution nonsense

* Fix auth endpoint configured for CI

---------

+++ lib-connect/Protocol/GuidGenerator.cs
@@ -102,7 +102,8 @@ public static class GuidGenerator
         var entropy = Random.Shared.Next(0, 65536);
 
         // Combine timestamp (48 bits) with entropy (16 bits)
-        return ((ulong)timestamp << 16) | (ulong)entropy;
+        // Cast entropy to uint first to avoid sign extension issues
+        return ((ulong)timestamp << 16) | (uint)entropy;
     }
 
     /// <summary>
